refactor: SceneFactory 삭제
